Avoid duplication through refactoring code in `maestro-xcuitest-driver`
- Copy the artifacts in xcuitest driver module
- Rename `Simctl` to `XCRunnerSimctl`
- Moved the XCUITestDriver in xcuitest driver module
