+++ maestro-utils/build.gradle
@@ -7,6 +7,8 @@ plugins {
 
 
 dependencies {
+    api 'com.squareup.okio:okio:3.0.0'
+
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.9.0'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.9.0'
     testImplementation "com.google.truth:truth:1.1.3"
