--- maestro-xcuitest-driver/src/main/kotlin/util/Simctl.kt
@@ -1,46 +1,0 @@
-package util
-
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-
-object Simctl {
-
-    fun listApps(): Set<String> {
-        val process = ProcessBuilder("bash", "-c", "xcrun simctl listapps booted | plutil -convert json - -o -").start()
-
-        val json = String(process.inputStream.readBytes())
-
-        val mapper = jacksonObjectMapper()
-        val appsMap = mapper.readValue(json, Map::class.java) as Map<String, Any>
-
-        return appsMap.keys
-    }
-
-    fun setProxy(host: String, port: Int) {
-        ProcessBuilder("networksetup", "-setwebproxy", "Wi-Fi", host, port.toString())
-            .redirectErrorStream(true)
-            .start()
-            .waitFor()
-        ProcessBuilder("networksetup", "-setsecurewebproxy", "Wi-Fi", host, port.toString())
-            .redirectErrorStream(true)
-            .start()
-            .waitFor()
-    }
-
-    fun resetProxy() {
-        ProcessBuilder("networksetup", "-setwebproxystate", "Wi-Fi", "off")
-            .redirectErrorStream(true)
-            .start()
-            .waitFor()
-        ProcessBuilder("networksetup", "-setsecurewebproxystate", "Wi-Fi", "off")
-            .redirectErrorStream(true)
-            .start()
-            .waitFor()
-    }
-
-    fun screenshot(path: String) {
-        CommandLineUtils.runCommand(
-            listOf("xcrun", "simctl", "io", "booted", "screenshot", path),
-            waitForCompletion = true
-        )
-    }
-}
